app/requirements.txt
strands-agents[openai]
strands-agents-tools
httpx
python-dotenv
pyyaml
ddgs
faster-whisper
croniter
yt-dlp
kokoro
soundfile
